+++ packages/button/src/Component.tsx
@@ -52,6 +52,11 @@ export type ComponentProps = {
      * Показать лоадер
      */
     loading?: boolean;
+
+    /**
+     * Не переносить текст кнопки на новую строку
+     */
+    nowrap?: boolean;
 };
 
 type AnchorButtonProps = ComponentProps & AnchorHTMLAttributes<HTMLAnchorElement>;
@@ -71,6 +76,7 @@ export const Button = React.forwardRef<HTMLAnchorElement | HTMLButtonElement, Bu
             dataTestId,
             href,
             loading = false,
+            nowrap = false,
             ...restProps
         },
         ref,
@@ -88,7 +94,8 @@ export const Button = React.forwardRef<HTMLAnchorElement | HTMLButtonElement, Bu
                     [styles.focused]: focused,
                     [styles.block]: block,
                     [styles.iconOnly]: !children,
-                    [styles.loading]: loading && !href,
+                    [styles.nowrap]: nowrap,
+                    [styles.loading]: loading,
                 },
                 className,
             ),
@@ -99,7 +106,7 @@ export const Button = React.forwardRef<HTMLAnchorElement | HTMLButtonElement, Bu
             <React.Fragment>
                 {leftAddons && <span className={cn(styles.addons)}>{leftAddons}</span>}
                 {children && <span className={cn(styles.text)}>{children}</span>}
-                {loading && !href && <Loader className={cn(styles.loader)} />}
+                {loading && <Loader className={cn(styles.loader)} />}
                 {rightAddons && <span className={cn(styles.addons)}>{rightAddons}</span>}
             </React.Fragment>
         );
@@ -143,4 +150,5 @@ Button.defaultProps = {
     size: 'm',
     block: false,
     loading: false,
+    nowrap: false,
 };